CP2 complete

+++ include/IOManagement.h
@@ -75,7 +75,7 @@ namespace ProdCon {
             std::cout << "    Sleep    ";
             std::cout << std::right << std::setw(6) << std::setfill(' ') << summary.at(4) << std::endl;
 
-            for (int i = 5; i < summary.size(); ++i) {
+            for (int i = 5; i < (int)summary.size(); ++i) {
                 std::cout << "    Thread";
                 std::cout << std::right << std::setw(3) << std::setfill(' ') << i - 4;
                 std::cout << std::right << std::setw(6) << std::setfill(' ') << summary.at(i) << std::endl;
